Delete job only it matches the driver and was not released previously

+++ src/Integrations/LaravelServiceProvider.php
@@ -25,7 +25,11 @@ class LaravelServiceProvider extends ServiceProvider
         ]);
 
         Queue::after(function (JobProcessed $event) {
-            $event->job->delete();
+            if ($event->job->getConnectionName() === 'sqs-plain') {
+                if (!$event->job->isReleased()) {
+                    $event->job->delete();
+                }
+            }
         });
     }
